+++ app/src/main/java/com/custom/transportation/base/BaseActivity.kt
@@ -1,0 +1,2 @@
+package com.custom.transportation.base
+
