+++ app/src/main/java/com/custom/transportation/base/BaseContract.kt
@@ -1,0 +1,7 @@
+package com.custom.transportation.base
+
+interface BaseContract {
+    interface View
+
+    interface Presenter
+}
