+++ app/src/main/java/com/custom/transportation/base/BaseRecyclerAdapter.kt
@@ -1,0 +1,2 @@
+package com.custom.transportation.base
+
