--- app/src/main/java/com/custom/transportation/model/BookmarkModel.kt
@@ -1,9 +1,0 @@
-package com.custom.transportation.model
-
-import com.custom.transportation.data.unit.BookmarkDatabase
-
-class BookmarkModel {
-
-    fun getBookmarkData() : List<Any> = BookmarkDatabase.getAll()
-
-}
