+++ app/src/main/java/com/custom/transportation/repository/model/BookmarkModel.kt
@@ -1,0 +1,9 @@
+package com.custom.transportation.repository.model
+
+import com.custom.transportation.repository.unit.BookmarkDatabase
+
+class BookmarkModel {
+
+    fun getBookmarkData() : List<Any> = BookmarkDatabase.getAll()
+
+}
