+++ app/src/main/java/com/custom/transportation/ui/activity/MainActivity.kt
@@ -1,11 +1,10 @@
-package com.custom.transportation
+package com.custom.transportation.ui.activity
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.viewpager.widget.ViewPager
+import com.custom.transportation.R
 import com.custom.transportation.ui.adapter.paper.SectionsPagerAdapter
-import com.custom.transportation.ui.common.MainTab
-import com.google.android.material.floatingactionbutton.FloatingActionButton
 import com.google.android.material.tabs.TabLayout
 
 class MainActivity : AppCompatActivity() {
