+++ app/src/main/java/com/custom/transportation/ui/adapter/recycler/BookmarkAdapter.kt
@@ -6,8 +6,8 @@ import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.custom.transportation.R
-import com.custom.transportation.data.unit.BusInfoData
-import com.custom.transportation.data.unit.BusStopData
+import com.custom.transportation.repository.unit.BusInfoData
+import com.custom.transportation.repository.unit.BusStopData
 
 class BookmarkAdapter : RecyclerView.Adapter<BookmarkAdapter.ViewHolder>() {
     private val items = mutableListOf<Any>()
