+++ app/src/main/java/com/custom/transportation/ui/adapter/recycler/BusStopAdapter.kt
@@ -6,13 +6,13 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
-import com.custom.transportation.BusStopDetailActivity
+import com.custom.transportation.ui.activity.BusStopDetailActivity
 import com.custom.transportation.R
-import com.custom.transportation.data.unit.BusStopData
-import com.custom.transportation.presenter.BusStop
-import com.custom.transportation.ui.common.IntentType
+import com.custom.transportation.repository.unit.BusStopData
+import com.custom.transportation.ui.contract.BusStopContract
+import com.custom.transportation.common.IntentType
 
-class BusStopAdapter(val presenter: BusStop.Presenter) : RecyclerView.Adapter<BusStopAdapter.ViewHolder>() {
+class BusStopAdapter(val presenter: BusStopContract.Presenter) : RecyclerView.Adapter<BusStopAdapter.ViewHolder>() {
     private val items = mutableListOf<BusStopData>()
 
     fun addItems(items : List<BusStopData>) {
