+++ app/src/main/java/com/custom/transportation/ui/contract/BookmarkContract.kt
@@ -1,0 +1,10 @@
+package com.custom.transportation.ui.contract
+
+import com.custom.transportation.repository.model.BookmarkModel
+
+interface BookmarkContract {
+
+    interface Presenter {
+        fun getBookmarkData() : List<Any>
+    }
+}
