+++ app/src/main/java/com/custom/transportation/ui/contract/BookmarkPresenter.kt
@@ -1,0 +1,10 @@
+package com.custom.transportation.ui.contract
+
+import com.custom.transportation.repository.model.BookmarkModel
+
+class BookmarkPresenter : BookmarkContract.Presenter {
+
+    private val model = BookmarkModel()
+
+    override fun getBookmarkData(): List<Any> = model.getBookmarkData()
+}
