+++ app/src/main/res/layout/activity_bus_stop.xml
@@ -6,7 +6,7 @@
     android:id="@+id/constraintLayout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BusStopDetailActivity">
+    tools:context=".ui.activity.BusStopDetailActivity">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/recycler_busstop"
